Twenty-eight of forty-eight: the first batch the loop was run at scale
Six sonnet agents over 48 leaf functions from the frontier, ranked by
callers and dealt round-robin so no agent drew only the hard end. Every
address landed a body, so all 48 leave the placeholder population:

  byte-exact claims   1269 -> 1297
  implemented         4112 -> 4160 pieces, 667858 -> 673837 bytes
  placeholder         1512 -> 1464 pieces

The twenty that did not match are landed with RULED-OUT lists rather than
discarded, which is the point of the grammar: a miss used to leave nothing
and the next agent restarted from zero. `--against-committed` now has an
incumbent to beat on each of them, and 0x005EABBF is MNEMONIC_ONLY - 66 of
66 mnemonics, operands only.

CLAIM AND MEASUREMENT AGREED ON ALL 48. No agent overclaimed. The four
that first measured REFUSED/NO_COMPILE were harness defects, fixed in the
commit that follows this one; their bodies were correct as written.

Two levers paid for themselves here and are now in the brief. Constructor
store order, four times - Font, Heap, Time, Spot - where the fields must
be assigned in the disassembly's write order, which a member-init list
CANNOT express because C++ runs initialisers in declaration order however
they are written; declaration order scores MNEMONIC_ONLY, which reads like
a hard register problem and is not one. And deleting cached locals, three
times - the original recomputes an expression at each use, and hoisting it
into a temp for readability forces an EBP frame and a stack spill.

Two catalogue findings, neither acted on here. `?set_modal@Time@@QAAXXZ`
and its three siblings are mangled QAA - public, non-static, __cdecl -
which cannot describe a function that reads its receiver from ECX
(0x00616860 stores it) nor three that have no receiver at all; same shape
as the 47 thunks corrected in bb41125f. And 0x005D2B20 takes a pointer in
esi under a convention VC6 will not emit for any C++ signature, with the
preceding function ending in `ret`, so it is not a fallthrough - a wall,
pending whether its two call sites are `call` or `jmp`.

Gate green: 1297 claims, 0 not reproduced.

+++ src/recovered/0064f09c.cpp
@@ -1,13 +1,11 @@
-// ORIGINAL: 0x0064F09C FILE
+// ORIGINAL: 0x0064F09C BYTE_EXACT FILE
+// working copy - scaffold materialised by --work
 // name      sub_64f09c
 // size      3 bytes
 // spans     0x0064F09C-0x0064F09F
 // prototype 
 // callers   2   call targets   0
-// kind      game
-// flags     hidden;sp_ready;purged_ok
 // calls     (none)
-// working copy - scaffold materialised by --work
 
 // GENERATED SKELETON - tools/emit_translation_unit.py
 // subject: sub_64f09c  at 0x0064F09C  (3 bytes)
@@ -42,11 +40,5 @@ typedef unsigned short uint16;
 typedef char int8;
 typedef unsigned char uint8;
 extern "C" int __cdecl sub_64f09c() {
-    // BODY GOES HERE.
-    //
-    // Reach fields by offset - the class is deliberately empty:
-    //     char *self = reinterpret_cast<char *>(this);
-    //     int v = *reinterpret_cast<int *>(self + 0x24);
-
-    return (int)0;  // PLACEHOLDER - replace with the body
+    return 0;
 }
